runtime command verification

+++ PokeStar/PokeStar/Modules/DexCommands.cs
@@ -13,42 +13,46 @@ namespace PokeStar.Modules
       [Command("dex")]
       public async Task Dex([Remainder] string text)
       {
-         var name = GetPokemon(text);
-         Pokemon pokemon = Connections.Instance().GetPokemon(name);
-         if (pokemon == null)
+         if (ChannelRegisterCommand.IsRegisteredChannel(Context.Guild.Id, Context.Channel.Id))
          {
-            EmbedBuilder embed = new EmbedBuilder();
-            embed.WithTitle("PokeDex Command Error");
-            embed.WithDescription($"Pokemon {name} cannot be found.");
-            embed.WithColor(Color.DarkRed);
-            await Context.Channel.SendMessageAsync(null, false, embed.Build()).ConfigureAwait(false);
+            var name = GetPokemon(text);
+            Pokemon pokemon = Connections.Instance().GetPokemon(name);
+            if (pokemon == null)
+            {
+               EmbedBuilder embed = new EmbedBuilder();
+               embed.WithTitle("PokeDex Command Error");
+               embed.WithDescription($"Pokemon {name} cannot be found.");
+               embed.WithColor(Color.DarkRed);
+               await Context.Channel.SendMessageAsync(null, false, embed.Build()).ConfigureAwait(false);
+            }
+            else
+            {
+               var fileName = Connections.GetPokemonPicture(pokemon.Name);
+               Connections.CopyFile(fileName);
+
+               EmbedBuilder embed = new EmbedBuilder();
+               embed.WithTitle($@"#{pokemon.Number} {pokemon.Name}");
+               embed.WithDescription(pokemon.Description);
+               embed.WithThumbnailUrl($"attachment://{fileName}");
+               embed.AddField("Type", pokemon.TypeToString(), true);
+               embed.AddField("Weather Boosts", pokemon.WeatherToString(), true);
+               embed.AddField("Details", pokemon.DetailsToString(), true);
+               embed.AddField("Stats", pokemon.StatsToString(), true);
+               embed.AddField("Resistances", pokemon.ResistanceToString(), true);
+               embed.AddField("Weaknesses", pokemon.WeaknessToString(), true);
+               embed.AddField("Fast Moves", pokemon.FastMoveToString(), true);
+               embed.AddField("Charge Moves", pokemon.ChargeMoveToString(), true);
+               embed.AddField("Counters", pokemon.CounterToString(), false);
+               embed.WithColor(Color.Red);
+               embed.WithFooter("* denotes STAB move ! denotes Legacy move");
+
+               await Context.Channel.SendFileAsync(fileName, embed: embed.Build()).ConfigureAwait(false);
+
+               Connections.DeleteFile(fileName);
+            }
          }
          else
-         {
-            var fileName = Connections.GetPokemonPicture(pokemon.Name);
-            Connections.CopyFile(fileName);
-
-            EmbedBuilder embed = new EmbedBuilder();
-            embed.WithTitle($@"#{pokemon.Number} {pokemon.Name}");
-            embed.WithDescription(pokemon.Description);
-            embed.WithThumbnailUrl($"attachment://{fileName}");
-            embed.AddField("Type", pokemon.TypeToString(), true);
-            embed.AddField("Weather Boosts", pokemon.WeatherToString(), true);
-            embed.AddField("Details", pokemon.DetailsToString(), true);
-            embed.AddField("Stats", pokemon.StatsToString(), true);
-            embed.AddField("Resistances", pokemon.ResistanceToString(), true);
-            embed.AddField("Weaknesses", pokemon.WeaknessToString(), true);
-            embed.AddField("Fast Moves", pokemon.FastMoveToString(), true);
-            embed.AddField("Charge Moves", pokemon.ChargeMoveToString(), true);
-            embed.AddField("Counters", pokemon.CounterToString(), false);
-            embed.WithColor(Color.Red);
-            embed.WithFooter("* denotes STAB move ! denotes Legacy move");
-
-            await Context.Channel.SendFileAsync(fileName, embed: embed.Build()).ConfigureAwait(false);
-
-            Connections.DeleteFile(fileName);
-         }
-
+            await Context.Channel.SendMessageAsync("This channel is not registered for commands.");
       }
       [Command("cp")]
       public async Task CP([Remainder] string text)
